Add files via upload
BITMANU has already earned the distinction of releasing the first ever line of 3nm ASIC miners, and BM Pro is the most powerful product in this range of miners. Equipped with the latest 3nm silicon semiconductor chips, it offers increased speed and transistor density, and reduced power consumption, resulting in significantly higher mining efficiency.

The hash rates with which BM Pro is able to mine Bitcoin, Litecoin, Ethereum, and Monero have not been heard of before its launch. As a result of its unrivalled hash power and power consumption, BM Pro has got the better of all currently available mining hardware in terms of profitability. On an average, all users of the product have been able to recover their initial investment with a month.